+++ lib/widgets/optionally_dismissible.dart
